Bugfix: fix audio_engine::compute_tick_size for non-integer bpm values

+++ src/core/include/hydrogen/audio_engine.h
@@ -147,7 +147,7 @@ public:
 	void unlock();
 	
 	
-	 static float compute_tick_size(int sampleRate, int bpm, int resolution);
+	 static float compute_tick_size(int sampleRate, float bpm, int resolution);
 
 	/** Returns #__sampler */
 	Sampler* get_sampler();
